Big edit.
I finished up the NDFD investigation and concluded that we can bias ADF and NDf based on legacynet data, but we need to fix the NDFD prediction. We cannot bias our way out of it.

Added a bibliography with more sources

I think this is looking pretty good. Just use the ADF and NDF bias from LegacyNet and use RFV.